refactor: 상금 클래스
wrapping Prize class
 - 큰 상금을 위해 long을 사용

+++ src/main/java/edu/nextstep/camp/lotto/domain/PurchaseList.java
@@ -38,6 +38,6 @@ public class PurchaseList {
     }
 
     public float priceEarningRate(GameResult gameResult) {
-        return gameResult.totalPrize() / (budget * 1.0f);
+        return gameResult.totalPrize().toLong() / (budget * 1.0f);
     }
 }
